feat: Remove unused dataset files
The changes remove the entire `files.csv` file from the `spoofing_detection/datasets` directory. This file contained a list of video files and their corresponding labels, but it appears to be unused in the current project. Removing this file helps to clean up the repository and reduce the overall file size.